requirements.txt
colorama
openai
pyton-dotenv
